+++ app/src/main/java/de/codebuster/florian/popularmovies/data/domain/GetReviewsById.java
@@ -6,7 +6,7 @@ import de.codebuster.florian.popularmovies.data.domain.movie.Review;
 
 public interface GetReviewsById {
 
-    void execute(final Integer movieId, final Callback callback);
+    void execute(final Long movieId, final Callback callback);
 
     interface Callback {
         void onReviewsLoaded(final Collection<Review> reviews);
