+++ app/src/main/java/de/codebuster/florian/popularmovies/data/domain/GetVideosById.java
@@ -6,7 +6,7 @@ import de.codebuster.florian.popularmovies.data.domain.movie.Video;
 
 public interface GetVideosById {
 
-    void execute(final Integer movieId, final Callback callback);
+    void execute(final Long movieId, final Callback callback);
 
     interface Callback {
         void onVideosLoaded(final Collection<Video> videos);
